pyproject.toml
[tool.poetry]
name = "udemy-langchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11.3"
gradio = "3.31.0"
slack-bolt = "1.18.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.12"
langchain-community = "^0.0.28"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
